handle errors and retry fetch

+++ src/util/chapter-cleanup.js
@@ -8,10 +8,14 @@ const l      = require('./log');
  * @param {!array} completedChapters Array of full paths to chapter directories.
  */
 module.exports = completedChapters => {
-  completedChapters.forEach(toDel => {
-    l.info(`rmdir ${toDel}`);
-    rimraf(toDel, () => {
-      l.debug(`nuked ${toDel}`);
+  if(!global.errors) {
+    completedChapters.forEach(toDel => {
+      l.info(`rmdir ${toDel}`);
+      rimraf(toDel, () => {
+        l.debug(`nuked ${toDel}`);
+      });
     });
-  });
+  } else {
+    l.warn(`@completedChapters - global.errors is set - NOT nuking source directories`);
+  }
 };
